+++ src/mainwindow.h
@@ -30,6 +30,13 @@ public:
         ReportViewTab
     };
 
+    enum EzPublishRemoteFileListColumns {
+        FileNameColumn = 0,
+        SizeColumn,
+        MTimeColumn,
+        DownloadColumn
+    };
+
     Q_ENUMS(ToolTabs)
 
     explicit MainWindow(QWidget *parent = 0);
@@ -193,8 +200,7 @@ private:
 
 class FileSizeTableWidgetItem : public QTableWidgetItem {
 public:
-    bool operator <(const QTableWidgetItem &other) const
-    {
+    bool operator <(const QTableWidgetItem &other) const {
         return data(Qt::UserRole).toInt() < other.data(Qt::UserRole).toInt();
     }
 };